Retire cozorocks C++ subcrate in favour of pure-Rust rocksdb
Replaces the custom cozorocks cxx FFI wrapper (and its 42 MB vendored
rocksdb submodule) with the community-maintained rust-rocksdb 0.22
crate, wired up as the single storage-rocksdb backend. The previously
experimental storage-new-rocksdb feature is folded in; "rocksdb" is
now the only RocksDB option and dispatches through the pure-Rust path.

Retires the cozorocks workspace member entirely: deletes cozorocks/,
the Facebook rocksdb submodule, the bridge/ C++ headers and .cpp
translation units, and the [profile.dev.package."cozorocks"] override.
Drops the cozorocks dependency from cozo-core and re-points jemalloc
and io-uring feature pass-throughs at rocksdb? instead.

RocksDB tuning that previously lived in cozorocks (9.9-bits-per-key
bloom filter with whole-key filtering, 9-byte prefix extractor matched
to KEY_PREFIX_LEN) is ported onto the pure-Rust backend via
BlockBasedOptions and SliceTransform::create_fixed_prefix. One
behavioural regression: external RocksDB options-file loading (the
"drop an OPTIONS-* file in your data dir" escape hatch) is not yet
wired up against rust-rocksdb's load_latest API; if an options file
is present the backend logs a warning and falls back to the built-in
tuning. README.md documents this as tracked future work.

Smoke benchmarks (time_travel, BASE=1000, MAX_K=100) against the
retired cozorocks build show the pure-Rust backend at rough parity:
inserts 1.8-2.3x faster (FFI-heavy write path), point reads +3 to
+7%, aggregations 20-29% slower on small datasets. The FFI-elimination
win on the write path is the dominant effect; aggregation deltas are
within noise at smoke scale and can be revisited via block-cache/
compression tuning if needed. All 172 cozo-core lib tests pass.

Motivation: cozorocks was a bespoke FFI subcrate only this repo
understood, gating any rocksdb upgrade on re-vetting hand-written
cxx bridge code, and the 42 MB vendored C++ rocksdb source was the
single slowest item in the build (~15 minutes cold) and the biggest
obstacle to cross-compilation. The pure-Rust rocksdb crate tracks
upstream directly and makes the build story for the whole workspace
"plain cargo build".
