+++ README.md
@@ -294,22 +294,16 @@ you can still try to compile your version to use, maybe with some tweaks in the 
 ### Tuning the RocksDB backend for CozoDB
 
 RocksDB has a lot of options, and by tuning them you can achieve better performance
-for your workload. This is probably unnecessary for 95% of users, but if you are the
-remaining 5%, CozoDB gives you the options to tune RocksDB directly if you are using the
-RocksDB storage engine.
+for your workload. CozoDB ships with sensible defaults (bloom filters, a prefix
+extractor matched to Cozo's key layout) which are probably fine for 95% of users.
 
-When you create the CozoDB instance with the RocksDB backend option, you are asked to
-provide a path to a directory to store the data (will be created if it does not exist).
-If you put a file named `options` inside this directory, the engine will expect this
-to be a [RocksDB options file](https://github.com/facebook/rocksdb/wiki/RocksDB-Options-File)
-and use it. If you are using the standalone `cozo` executable, you will get a log message if
-this feature is activated.
-
-Note that improperly set options can make your database misbehave!
-In general, you should run your database once, copy the options file from `data/OPTIONS-XXXXXX`
-from within your database directory, and use that as a base for your customization.
-If you are not an expert on RocksDB, we suggest you limit your changes to adjusting those numerical
-options that you at least have a vague understanding.
+Note: earlier releases of CozoDB supported dropping a RocksDB options file inside
+the database directory to override engine configuration. This feature relied on
+the bespoke `cozorocks` C++ FFI subcrate, which has been retired in favour of the
+pure-Rust `rocksdb` crate; external options-file loading is not currently wired up
+against the new backend. If CozoDB finds an `options` file in your data directory
+it will log a warning and fall back to the built-in tuning. Re-enabling this as
+a feature against the pure-Rust backend is tracked as future work.
 
 ## Architecture
