+++ cozo-core/Cargo.toml
@@ -32,16 +32,15 @@ storage-sqlite-src = ["dep:sqlite3-src", "sqlite3-src/bundled"]
 ## RocksDB is hard to compile on some platforms, uses more resources than SQLite,
 ## but is very performant and supports an extremely high level of concurrency.
 ## You can also [fine-tune](https://github.com/cozodb/cozo/blob/main/TUNING_ROCKSDB.md) RocksDB options.
-storage-rocksdb = ["dep:cozorocks"]
-storage-new-rocksdb = ["dep:rocksdb"]
+storage-rocksdb = ["dep:rocksdb"]
 ## Enables the graph algorithms.
 graph-algo = ["graph", "rayon"]
 ## Allows the utilities to make web requests to fetch data.
 requests = ["dep:minreq"]
 ## Uses jemalloc as the global allocator, can make a difference in performance.
-jemalloc = ["dep:tikv-jemallocator-global", "cozorocks?/jemalloc"]
+jemalloc = ["dep:tikv-jemallocator-global", "rocksdb?/jemalloc"]
 ## Enables io-uring option for the RocksDB storage
-io-uring = ["cozorocks?/io-uring"]
+io-uring = ["rocksdb?/io-uring"]
 ## Polyfills for the WASM target
 wasm = ["uuid/js", "dep:js-sys"]
 
@@ -120,7 +119,6 @@ document-features = "0.2.8"
 rayon = { version = "1.10.0", optional = true }
 minreq = { version = "2.11.2", features = ["https-rustls"], optional = true }
 tikv-jemallocator-global = { version = "0.5.0", optional = true }
-cozorocks = { path = "../cozorocks", version = "0.1.7", optional = true }
 rocksdb = { version = "0.22.0", optional = true }
 sled = { version = "0.34.7", optional = true }
 tikv-client = { version = "0.3.0", optional = true }
