+++ cozo-core/src/lib.rs
@@ -60,8 +60,6 @@ pub use runtime::temp_store::RegularTempStore;
 pub use storage::mem::{new_cozo_mem, MemStorage};
 #[cfg(feature = "storage-rocksdb")]
 pub use storage::rocks::{new_cozo_rocksdb, RocksDbStorage};
-#[cfg(feature = "storage-new-rocksdb")]
-pub use storage::newrocks::{new_cozo_newrocksdb, NewRocksDbStorage};
 #[cfg(feature = "storage-sled")]
 pub use storage::sled::{new_cozo_sled, SledStorage};
 #[cfg(feature = "storage-sqlite")]
@@ -112,9 +110,6 @@ pub enum DbInstance {
     #[cfg(feature = "storage-rocksdb")]
     /// RocksDB storage
     RocksDb(Db<RocksDbStorage>),
-    #[cfg(feature = "storage-new-rocksdb")]
-    /// New RocksDB storage
-    NewRocksDb(Db<NewRocksDbStorage>),
     #[cfg(feature = "storage-sled")]
     /// Sled storage (experimental)
     Sled(Db<SledStorage>),
@@ -136,7 +131,6 @@ impl DbInstance {
     /// * `mem`
     /// * `sqlite`
     /// * `rocksdb`
-    /// * `newrocksdb`
     /// * `sled`
     /// * `tikv`
     ///
@@ -154,8 +148,6 @@ impl DbInstance {
             "sqlite" => Self::Sqlite(new_cozo_sqlite(path)?),
             #[cfg(feature = "storage-rocksdb")]
             "rocksdb" => Self::RocksDb(new_cozo_rocksdb(path)?),
-            #[cfg(feature = "storage-new-rocksdb")]
-            "newrocksdb" => Self::NewRocksDb(new_cozo_newrocksdb(path)?),
             #[cfg(feature = "storage-sled")]
             "sled" => Self::Sled(new_cozo_sled(path)?),
             #[cfg(feature = "storage-tikv")]
@@ -191,8 +183,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.get_fixed_rules(),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.get_fixed_rules(),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.get_fixed_rules(),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.get_fixed_rules(),
             #[cfg(feature = "storage-tikv")]
@@ -230,8 +220,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.run_script_ast(payload, cur_vld, mutability),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.run_script_ast(payload, cur_vld, mutability),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.run_script_ast(payload, cur_vld, mutability),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.run_script_ast(payload, cur_vld, mutability),
             #[cfg(feature = "storage-tikv")]
@@ -305,8 +293,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.export_relations(relations),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.export_relations(relations),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.export_relations(relations),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.export_relations(relations),
             #[cfg(feature = "storage-tikv")]
@@ -347,8 +333,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.import_relations(data),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.import_relations(data),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.import_relations(data),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.import_relations(data),
             #[cfg(feature = "storage-tikv")]
@@ -390,8 +374,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.backup_db(out_file),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.backup_db(out_file),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.backup_db(out_file),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.backup_db(out_file),
             #[cfg(feature = "storage-tikv")]
@@ -414,8 +396,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.restore_backup(in_file),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.restore_backup(in_file),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.restore_backup(in_file),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.restore_backup(in_file),
             #[cfg(feature = "storage-tikv")]
@@ -442,8 +422,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.import_from_backup(in_file, relations),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.import_from_backup(in_file, relations),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.import_from_backup(in_file, relations),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.import_from_backup(in_file, relations),
             #[cfg(feature = "storage-tikv")]
@@ -482,8 +460,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.register_callback(relation, capacity),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.register_callback(relation, capacity),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.register_callback(relation, capacity),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.register_callback(relation, capacity),
             #[cfg(feature = "storage-tikv")]
@@ -500,8 +476,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.unregister_callback(id),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.unregister_callback(id),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.unregister_callback(id),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.unregister_callback(id),
             #[cfg(feature = "storage-tikv")]
@@ -519,8 +493,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.register_fixed_rule(name, rule_impl),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.register_fixed_rule(name, rule_impl),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.register_fixed_rule(name, rule_impl),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.register_fixed_rule(name, rule_impl),
             #[cfg(feature = "storage-tikv")]
@@ -535,8 +507,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.unregister_fixed_rule(name),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.unregister_fixed_rule(name),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.unregister_fixed_rule(name),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.unregister_fixed_rule(name),
             #[cfg(feature = "storage-tikv")]
@@ -557,8 +527,6 @@ impl DbInstance {
             DbInstance::Sqlite(db) => db.run_multi_transaction(write, payloads, results),
             #[cfg(feature = "storage-rocksdb")]
             DbInstance::RocksDb(db) => db.run_multi_transaction(write, payloads, results),
-            #[cfg(feature = "storage-new-rocksdb")]
-            DbInstance::NewRocksDb(db) => db.run_multi_transaction(write, payloads, results),
             #[cfg(feature = "storage-sled")]
             DbInstance::Sled(db) => db.run_multi_transaction(write, payloads, results),
             #[cfg(feature = "storage-tikv")]
