--- cozorocks/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "cozorocks"
-version = "0.1.7"
-edition = "2021"
-license = "MPL-2.0"
-authors = ["Ziyang Hu"]
-description = "Bindings to RocksDB used by the cozo crate"
-homepage = "https://www.cozodb.org"
-repository = "https://github.com/cozodb/cozo"
-documentation = "https://docs.cozodb.org"
-exclude = ["rocksdb/docs", "rocksdb/java"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-
-[dependencies]
-cxx = "1.0.69"
-miette = "5.10.0"
-libc = "0.2"
-tikv-jemalloc-sys = { version = "0.5", features = ["unprefixed_malloc_on_supported_platforms"], optional = true }
-lz4-sys = { version = "1.9" }
-zstd-sys = { version = "2.0", features = ["zdict_builder"] }
-
-[features]
-jemalloc = ["tikv-jemalloc-sys"]
-io-uring = ["pkg-config"]
-
-[build-dependencies]
-cxx-build = "1.0.69"
-pkg-config = { version = "0.3.25", optional = true }
-cc = { version = "1.0", features = ["parallel"] }
