--- cozorocks/README-zh.md
@@ -1,3 +1,0 @@
-# Cozorocks
-
-对 RocksDB C++ 接口的封装。
