--- cozorocks/README.md
@@ -1,3 +1,0 @@
-# Cozorocks
-
-Bindings to RocksDB's C++ API.
