--- cozorocks/bridge/bridge.h
@@ -1,17 +1,0 @@
-// Copyright 2022, The Cozo Project Authors.
-//
-// This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
-// If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-#ifndef COZOROCKS_BRIDGE_H
-#define COZOROCKS_BRIDGE_H
-
-#include "db.h"
-#include "slice.h"
-#include "tx.h"
-#include "status.h"
-#include "opts.h"
-#include "iter.h"
-
-#endif //COZOROCKS_BRIDGE_H
